Separerer modal i to komponenter
